+++ tauro/blueprints/api_oauth2_v1/endpoints/consultar_turnos.py
@@ -34,7 +34,7 @@ class ConsultarTurnos(Resource):
         turnos = (
             Turno.query.join(TurnoEstado)
             .join(TurnoTipo)
-            .filter(TurnoEstado.nombre != "ATENDIENDO", TurnoEstado.nombre != "COMPLETADO", TurnoEstado.nombre != "CANCELADO")
+            .filter(or_(TurnoEstado.nombre == "EN ESPERA", TurnoEstado.nombre != "PASE A VENTANILLA"))
             .filter(Turno.estatus == "A")
             .order_by(
                 # 1. Prioridad por estado: EN ESPERA primero (valor 0), el resto después (valor 1)
@@ -72,8 +72,7 @@ class ConsultarTurnos(Resource):
             .join(TurnoTipo)
             .filter(
                 or_(
-                    TurnoEstado.nombre == "EN ESPERA",
-                    TurnoEstado.nombre == "PASE A VENTANILLA",
+                    TurnoEstado.nombre == "ATENDIENDO",
                     TurnoEstado.nombre == "ATENDIENDO EN CUBICULO",
                 )
             )